tests/submission/MOSQUITO/event-source/event-type-eventsource.html

<!DOCTYPE html>
<html>
<head>
	<title>When the EventSource() constructor is invoked, the UA must run these steps: Create a new EventSource object:</title>
  <meta charset="utf-8" />
  	<meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
	  <link rel="author" title="Fabrice Clari" href="mailto:f.clari@inno-group.com">
	  <link rel="author" title="Dimitri Bocquet" href="mailto:Dimitri.Bocquet@mosquito-fp7.eu">
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="/resources/WebIDLParser.js"></script>
	<script src="/resources/idlharness.js"></script>
</head>
<body>
	<h1>When the EventSource() constructor is invoked, the UA must run these steps: Create a new EventSource object:</h1>
	<div id="log"></div>
  <script>
    var source = new EventSource('http://html5.mosquito-fp7.eu/messages/t0002');
    test(function(){assert_true((source.constructor.toString().indexOf("EventSource") != -1))},"The object is EventSource.");
  </script>
  </div>
</body>
</html>
